Add files via upload
Coin_Counter_Verison_2 is the second version of the script currently used in the Blue Box AC Controller. (Latest is 11, I think)
It's actually not that far off form the latest version. Most of the revisions were used to figure out that we needed to use a analog Low Pass Filter in order to handle the inrush current EMI when the AC turns on.